+++ tensorcircuit/cons.py
@@ -34,6 +34,8 @@ modules = [
     "tensorcircuit.quantum",
     "tensorcircuit.simplify",
     "tensorcircuit.interfaces",
+    "tensorcircuit.experimental",
+    "tensorcircuit.utils",
     "tensorcircuit.backends",
     "tensorcircuit.backends.abstract_backend",
     "tensorcircuit.backends.backend_factory",